requirements.txt
streamlit
pandas
matplotlib
pdfplumber
openpyxl
google-generativeai
plotly